requirements.txt
flask==2.2.3
pandas==1.5.3
numpy==1.26.4
openai==1.3.0
python-dotenv==1.0.0
werkzeug==2.2.3
gunicorn==20.1.0
